Move BSTRecord class definition into DataFormats/Scalers, and cleanup.
