+++ DQMServices/Components/BuildFile.xml
@@ -1,10 +1,11 @@
 <use   name="CondFormats/DataRecord"/>
 <use   name="CondFormats/RunInfo"/>
-<use   name="DataFormats/Scalers"/>
 <use   name="DataFormats/L1GlobalTrigger"/>
+<use   name="DataFormats/Provenance"/>
+<use   name="DataFormats/Scalers"/>
 <use   name="DQMServices/ClientConfig"/>
-<use   name="EventFilter/Utilities"/>
 <use   name="DQMServices/Core"/>
+<use   name="EventFilter/Utilities"/>
 <use   name="FWCore/Catalog"/>
 <use   name="FWCore/Framework"/>
 <use   name="FWCore/ParameterSet"/>
