+++ DataFormats/FEDRawData/src/classes_def.xml
@@ -16,9 +16,6 @@
   <version ClassVersion="11" checksum="2009823250"/>
   <version ClassVersion="10" checksum="120358665"/>
  </class>
- <class name="BSTRecord" ClassVersion="3">
-  <version ClassVersion="3" checksum="3262583400"/>
- </class>
  <class name="fedh_struct" ClassVersion="10">
   <version ClassVersion="10" checksum="3158059546"/>
  </class>
@@ -27,5 +24,4 @@
  </class>
  <class name="edm::Wrapper<FEDRawDataCollection>" splitLevel="0"/>
  <class name="edm::RefProd<FEDRawDataCollection>"/>
- <class name="edm::Wrapper<BSTRecord>"/>
 </lcgdict>
