+++ DataFormats/Scalers/src/classes.h
@@ -10,6 +10,7 @@
 #include "DataFormats/Scalers/interface/DcsStatus.h"
 #include "DataFormats/Scalers/interface/BeamSpotOnline.h"
 #include "DataFormats/Scalers/interface/TimeSpec.h"
+#include "DataFormats/Scalers/interface/BSTRecord.h"
 #include "DataFormats/Common/interface/Wrapper.h"
 #include "DataFormats/Common/interface/Ref.h"
 #include "DataFormats/Common/interface/RefProd.h"
@@ -24,6 +25,7 @@ namespace DataFormats_Scalers {
     LumiScalers lumiScalers;
     DcsStatus dcsStatus;
     BeamSpotOnline beamSpotOnline;
+    BSTRecord bstRecord;
 
     edm::Wrapper<L1AcceptBunchCrossing> w_l1AcceptBunchCrossing;
     edm::Wrapper<L1TriggerScalers> w_l1TriggerScalers;
@@ -33,6 +35,7 @@ namespace DataFormats_Scalers {
     edm::Wrapper<LumiScalers> w_lumiScalers;
     edm::Wrapper<DcsStatus> w_dcsStatus;
     edm::Wrapper<BeamSpotOnline> w_beamSpotOnline;
+    edm::Wrapper<BSTRecord> w_bstRecord;
 
     edm::RefProd<L1AcceptBunchCrossing> l1AcceptBunchCrossingRef ;
     edm::RefProd<L1TriggerScalers> l1TriggerScalersRef ;
@@ -42,6 +45,7 @@ namespace DataFormats_Scalers {
     edm::RefProd<LumiScalers> lumiScalersRef ;
     edm::RefProd<DcsStatus> dcsStatusRef ;
     edm::RefProd<BeamSpotOnline> beamSpotOnlineRef ;
+    edm::RefProd<BSTRecord> bstRecordRef ;
 
     L1AcceptBunchCrossingCollection l1AcceptBunchCrossingCollection;
     edm::Wrapper<L1AcceptBunchCrossingCollection>
