+++ EventFilter/Utilities/plugins/BuildFile.xml
@@ -1,14 +1,15 @@
 <library   file="*.cc" name="EventFilterUtilitiesPlugins">
+  <use   name="DataFormats/FEDRawData"/>
+  <use   name="DataFormats/Scalers"/>
+  <use   name="EventFilter/Utilities"/>
+  <use   name="FWCore/Framework"/>
   <use   name="FWCore/MessageLogger"/>
   <use   name="FWCore/ServiceRegistry"/>
-  <use   name="FWCore/Framework"/>
   <use   name="FWCore/Sources"/>
   <use   name="FWCore/Utilities"/>
   <use   name="IOPool/Streamer"/>
-  <use   name="DataFormats/FEDRawData"/>
-  <use   name="EventFilter/Utilities"/>
-  <use   name="root"/>
-  <use   name="boost"/>
   <use   name="CLHEP"/>
+  <use   name="boost"/>
+  <use   name="root"/>
   <flags   EDM_PLUGIN="1"/>
 </library>
